Moves and cleans up pagination for courses
