--- app/templates/rounds/index.hbs
@@ -1,96 +1,0 @@
-<h2 class="text-center">
-  Rounds
-</h2>
-{{#if model}}
-  <table id="rounds" class="table table-striped">
-    <thead>
-      <tr>
-        <th {{action "sortBy" "courseName" on="click"}}>
-          Course
-          <i {{bind-attr class="sortCourseName"}}>
-          </i>
-        </th>
-        <th {{action "sortBy" "createdAt" on="click"}}>
-          Date Played
-          <i {{bind-attr class="sortCreatedAt"}}>
-          </i>
-        </th>
-        <th {{action "sortBy" "totalScore" on="click"}}>
-          Score
-          <i {{bind-attr class="sortTotalScore"}}>
-          </i>
-        </th>
-        <th>
-        </th>
-        <th>
-        </th>
-      </tr>
-    </thead>
-    <tbody>
-      {{#each scorecard in pagedContent}}
-        <tr>
-          <td>
-            {{scorecard.courseName}}
-          </td>
-          <td>
-            {{display-date scorecard.createdAt}}
-          </td>
-          <td>
-            {{scorecard.formattedTotals}}
-          </td>
-          <td>
-            {{#if scorecard.isFinished}}
-              {{#link-to "round" scorecard.round}}
-                Recap
-              {{/link-to}}
-            {{else}}
-              {{#link-to "turns" scorecard.round scorecard.nextUnplayedTurn.holeNumber}}
-                Jump to Hole
-                {{scorecard.nextUnplayedTurn.holeNumber}}
-              {{/link-to}}
-            {{/if}}
-          </td>
-          <td>
-            {{#unless scorecard.round.isStarted}}
-              <a {{action "deleteRound" scorecard.round on="click"}}>
-                Delete Round
-              </a>
-            {{/unless}}
-          </td>
-        </tr>
-      {{/each}}
-    </tbody>
-    <tfoot>
-      <tr>
-        <td colspan="5">
-          {{#unless isFirstPage}}
-            <div class="pull-left">
-              <a {{action "previousPage" on="click"}} class="btn btn-success btn-lg">
-                <span class="glyphicon glyphicon-chevron-left">
-                </span>
-              </a>
-            </div>
-          {{/unless}}
-          {{#unless isLastPage}}
-            <div class="pull-right">
-              <a {{action "nextPage" on="click"}} class="btn btn-success btn-lg">
-                <span class="glyphicon glyphicon-chevron-right">
-                </span>
-              </a>
-            </div>
-          {{/unless}}
-          <div class="text-center">
-            Page
-            {{page}}
-             of
-            {{totalPages}}
-          </div>
-        </td>
-      </tr>
-    </tfoot>
-  </table>
-{{else}}
-  <p>
-    You haven't played any rounds
-  </p>
-{{/if}}
